refactor: move script-execution related code to separate directory
