+++ src/script-execution/types.ts
@@ -1,0 +1,9 @@
+import { GoogleSpreadsheet } from 'google-spreadsheet';
+
+import { ScriptError } from '../script-error';
+
+export type ScriptExecutor = (
+  spreadsheet: GoogleSpreadsheet,
+) => Promise<string>;
+
+export type ScriptErrorHandler = (scriptError: ScriptError) => string;
